requirements.txt
aiosignal==1.2.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
loguru==0.6.0
Mako==1.2.2
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.21.0
random-user-agent==1.0.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
yarl==1.8.1
